Create a Simple Lambda Function
Lambda function created is used to replay messages from an SQS queue
to another lambda function.

This code will run as an AWS Lambda by receiving messages from an SQS messages
with a long poll and use each message as a payload when we invoke another
lambda.

These days, microservices architecture is widely used and instead of using one
big code/system we use a lot of little small components that talk to each other
via integration system such as a queueing mechanism.

Despite its flexibility, more small systems mean maintainability will be harder
and we might break stuff or at least one of the systems will and the best
practice around that is to dead-letter failed messages/data.